budget_api/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.3
django-debug-toolbar==5.2.0
django-extensions==4.1
django-ninja==1.4.3
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-decouple==3.8
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
